# Optimize read performance