This change results in a substantial performance boost for PTY read operations.

# Performance Comparison: Non-Blocking Read Test

| Metric                         | Before Optimization | After Optimization |
| ------------------------------ | ------------------- | ------------------ |
| **Total time**                 | 106.9877 ms         | 15.4 µs            |
| **Average time per read**      | 106.9877 ms         | 0.0077 ms          |
| **Total bytes read**           | 76                  | 76                 |