@shadow3aaa 2025-03-27
this closes #92