@andfoy 2025-03-31
I'm going to test these changes against `pywinpty` and see how they behave there, if everything goes fine, I'll merge this PR