@shadow3aaa 2025-04-29
> @shadow3aaa, sorry for the delay! This week I'll take a look into it!

Hi @andfoy , I carefully reviewed your discussion in #75 and the architecture design in base.rs. This leads me to think that the design of cache_thread seems unnecessary? I'll provide some explanation below; please correct me if I'm wrong. Thanks.

Currently, the process for reading process output on the main branch is as follows:

Start process Start reading thread Get process handle Loop to read output Send output to pipe Start cache thread Loop to read output from the reading thread Cache to cache Check for read requests If there is a read request, send the result to the main thread through another pipe 

I noticed that in #75, you mentioned:

> Stopping the cache could be improved by distinguishing the read from the stop message requests. However, the main issue lies in the main reading process, which can get I/O blocked by the process output, and locking on the stop message leads to hunger issues, since the continuous read would not take place whilst the process is waiting for the stop signal.

But, from my understanding, the cache does not seem to help solve this issue. Meanwhile, the only function the cache provides appears to be allowing us to request a certain amount of data in characters instead of just reading one piece of pipe information at a time, essentially acting as a buffer. This adds an extra layer of abstraction, and I don't think it should be designed within PTYProcess, because the same functionality can be easily added externally via a buffer. If it’s implemented internally, it would make it harder to choose other solutions (of course, we could also send characters one by one directly from the reading thread, but that would be less efficient).

Also, regarding the termination of the reading thread, I noticed that blocking eventually occurs here:

```rust
if ReadFile(stream, Some(&mut buf_vec[..]), chars_read_mut, None).is_ok() { S_OK 
```

Currently, to allow the thread to exit, a channel is set up to pass a stop flag. However, if the read is blocked continuously and the process hasn’t finished, it seems that the thread cannot exit?

Actually, there is a Win32 API that can directly interrupt the waiting of ReadFile:

```
BOOL WINAPI CancelSynchronousIo( _In_ HANDLE hThread ); 
```

So I think we could call it in drop to interrupt the reading thread’s wait, and then use a simple Arc<AtomicBool> to check whether the thread should exit.