@andfoy 2025-05-01
Thanks for the detailed analysis @shadow3aaa! Let me answer some of the main discussion points:

> Meanwhile, the only function the cache provides appears to be allowing us to request a certain amount of data in characters instead of just reading one piece of pipe information at a time, essentially acting as a buffer.

This is a correct assessment, however, it is worth to mention that having the cache as facade also has the benefit of serializing the access to the actual reading handle, preventing contention if multiple threads were to emit reading requests.

> I noticed that blocking eventually occurs here:

According to https://learn.microsoft.com/en-us/windows/win32/api/namedpipeapi/nf-namedpipeapi-peeknamedpipe#remarks, synchronous I/O pipes can deadlock if multiple operations originating from multiple threads take place, which is the case regarding the continuous reading thread and checking if the process is still alive (which usually takes place from a different thread). As I mentioned on https://github.com/andfoy/pywinpty/issues/463#issuecomment-2450162112, the straight forward solution would be to enable asynchronous I/O instead.

Something alongside these lines: https://github.com/microsoft/terminal/blob/08e76da3a1687a1303733406b684a9a6ab523acb/src/cascadia/TerminalConnection/ConptyConnection.cpp#L668

However, this would imply somehow having to compile and link the latest ConPTY version? I'll need to take a more detailed look at it

> So I think we could call it in drop to interrupt the reading thread’s wait, and then use a simple Arc to check whether the thread should exit.

I think the issue is that doing so would cause reading to be non-blocking? The ideal solution would be having the Arc + Async I/O