@ofek 2025-05-01
> if the read is blocked continuously and the process hasn’t finished, it seems that the thread cannot exit

Perhaps that is the source of my issue (and described [here](https://github.com/andfoy/pywinpty/issues/423#issuecomment-2323083529) by someone else)?

I was doing a blocking read of the `PTY` object provided by https://github.com/andfoy/pywinpty in a separate thread and catching ctrl+c. When this happens I attempt to terminate the process by doing `os.kill(pty.pid, signal.SIGTERM)` but the read continues to block and the process stays up, seemingly never getting the signal.